switch to using load-history to find ert tests
This had to be done because feature symbols no longer exists.  So far all tests
created from ERT macros appear to be caught by this selector.

The other option was to use the project symbol prefix, but this could fail if
the prefix is an initialism or the test is named incorrectly.  Using load
history should be more robust.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>