Generated readme and texinfo manual
The documentation was moved to /doc and broken up so that there is now a manual
and the README includes sections of the manual.  The contributing guide was
mostly left as-is.

Table of contents no longer relies on org-toc

Several helper commands to export and view the correct files were created.

ERK now reads its own attributes to decide how to perform author renaming and
also uses these during document generation

The documentation generation was documented.  Oh god.

Lots of little documentation tweaks were done to format or support the exports.

generated documentation found in new README.md, CONTRIBUTING.md, and manual.texi

While considering how this will work for generic template repositories, it
became clear that looking at ERK itself to figure out what to replace is not
adding value, but it is adding overhead

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>