Support for external templates
ERK is no longer a good demonstration project because it's getting too big.
This also makes string replacement hard because if your derivative needs to
retain instructions about using ERK, the clones get strings replaced that were
needed for the instructions.

Implementation was done by creating a template system and pointing it towards
erk-basic, a stripped down ERK project with just one command and one test.

Licenses are a bit annoying in this release.  I want to go ahead and switch to a
proper license package and take the licenses out of the templates.

For now, the license had to be added to the erk-basic in order to maintain rough
parity.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>