! use root-relative links
This is pretty incorrect.  More fancy solutions will do some translation on the
links to fix the offsets.  Instead of choosing a fancy solution now, the links
are made to work in the outputs, sacrificing their correctness in the inputs.

This commit will likely be rolled back after a more sustainable solution for
link translation is adopted.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>